test: Split API v2 tests
Duplication here seems better than weird and hard to understand
parameterized version.
